SOLR-14381: Grouping to use long instead of int to avoid overflows

+++ solr/core/src/java/org/apache/solr/request/SimpleFacets.java
@@ -341,7 +341,7 @@ public class SimpleFacets {
         .add(mainQueryFilter, Occur.FILTER)
         .build();
     searcher.search(filteredFacetQuery, collector);
-    return collector.getGroupCount();
+    return (int) collector.getGroupCount();
   }
 
   enum FacetMethod {
